+++ src/woocommerce_hipayenterprise/includes/gateways/class-hipay-gateway-abstract.php
@@ -64,6 +64,26 @@ class Hipay_Gateway_Abstract extends WC_Payment_Gateway
     protected $paymentProduct;
 
     /**
+     * @var
+     */
+    protected $username;
+
+    /**
+     * @var
+     */
+    protected $password;
+
+    /**
+     * @var
+     */
+    protected $sandbox;
+
+    /**
+     * @var Hipay_Available_Payment
+     */
+    protected $availablePayment;
+
+    /**
      * Hipay_Gateway_Abstract constructor.
      */
     public function __construct()
@@ -110,28 +130,41 @@ class Hipay_Gateway_Abstract extends WC_Payment_Gateway
             true
         );
 
-        $sandbox = $this->confHelper->getAccount()["global"]["sandbox_mode"];
-        $username = ($sandbox) ? $this->confHelper->getAccount()["sandbox"]["api_tokenjs_username_sandbox"]
+        $this->sandbox = $this->confHelper->getAccount()["global"]["sandbox_mode"];
+        $this->username = ($this->sandbox) ? $this->confHelper->getAccount()["sandbox"]["api_tokenjs_username_sandbox"]
             : $this->confHelper->getAccount()["production"]["api_tokenjs_username_production"];
-        $password = ($sandbox) ? $this->confHelper->getAccount()["sandbox"]["api_tokenjs_password_publickey_sandbox"]
+        $this->password = ($this->sandbox) ? $this->confHelper->getAccount()["sandbox"]["api_tokenjs_password_publickey_sandbox"]
             : $this->confHelper->getAccount()["production"]["api_tokenjs_password_publickey_production"];
+
+
 
         wp_localize_script(
             'hipay-js-front',
             'hipay_config',
             array(
-                "apiUsernameTokenJs" => $username,
-                "apiPasswordTokenJs" => $password,
+                "apiUsernameTokenJs" => $this->username,
+                "apiPasswordTokenJs" => $this->password,
                 "lang" => substr(get_locale(), 0, 2),
-                "environment" => $sandbox ? "stage" : "production",
+                "environment" => $this->sandbox ? "stage" : "production",
                 "fontFamily" => $this->confHelper->getHostedFieldsStyle()["fontFamily"],
                 "color" => $this->confHelper->getHostedFieldsStyle()["color"],
                 "fontSize" => $this->confHelper->getHostedFieldsStyle()["fontSize"],
                 "fontWeight" => $this->confHelper->getHostedFieldsStyle()["fontWeight"],
                 "placeholderColor" => $this->confHelper->getHostedFieldsStyle()["placeholderColor"],
                 "caretColor" => $this->confHelper->getHostedFieldsStyle()["caretColor"],
-                "iconColor" => $this->confHelper->getHostedFieldsStyle()["iconColor"],
+                "iconColor" => $this->confHelper->getHostedFieldsStyle()["iconColor"]
             )
+        );
+
+        $paypalOptions = Hipay_Available_Payment::getInstance($this->confHelper)
+            ->getAvailablePaymentProducts('paypal')[0]['options'] ?? [];
+
+        wp_localize_script(
+            'hipay-js-front',
+            'paypal_version',
+            ['v2' => !empty($paypalOptions['provider_architecture_version'])
+                && $paypalOptions['provider_architecture_version'] === 'v1'
+                && !empty($paypalOptions['payer_id'])]
         );
     }