+++ src/woocommerce_hipayenterprise/includes/gateways/class-hipay-gateway-local-abstract.php
@@ -40,10 +40,8 @@ class Hipay_Gateway_Local_Abstract extends Hipay_Gateway_Abstract
             $this->title = $methodConf["displayName"]['en'];
         }
 
-        if (isset($methodConf["logo"])) {
-            if (file_exists(WC_HIPAYENTERPRISE_PATH_ASSETS . "/local_payments_images/" . $methodConf["logo"])) {
-                $this->icon = WC_HIPAYENTERPRISE_URL_ASSETS . "/local_payments_images/" . $methodConf["logo"];
-            }
+        if (isset($methodConf["logo"]) && file_exists(WC_HIPAYENTERPRISE_PATH_ASSETS . "local_payments_images/" . $methodConf["logo"])) {
+            $this->icon = WC_HIPAYENTERPRISE_URL_ASSETS . "local_payments_images/" . $methodConf["logo"];
         }
 
         $this->init_form_fields();
@@ -60,11 +58,15 @@ class Hipay_Gateway_Local_Abstract extends Hipay_Gateway_Abstract
         if ($methodConf["canRefund"]) {
             $this->supports[] = "refunds";
         }
+
+        if (isset($_GET['section']) && preg_match("/^hipayenterprise_paypal/", $_GET['section'])) {
+            $this->availablePayment = Hipay_Available_Payment::getInstance($this->confHelper);
+        }
     }
 
     public function isAvailable()
     {
-        return ('yes' === $this->enabled);
+        return 'yes' === $this->enabled;
     }
 
     public function payment_fields()
@@ -116,10 +118,11 @@ class Hipay_Gateway_Local_Abstract extends Hipay_Gateway_Abstract
         $this->process_template(
             'admin-paymentlocal-settings.php',
             'admin',
-            array(
+            [
                 'configurationPaymentMethod' => $this->confHelper->getLocalPayment($this->paymentProduct),
+                'isPayPalV2' => $this->isPaypalV2(),
                 'method' => $this->paymentProduct
-            )
+            ]
         );
 
         return ob_get_clean();
@@ -138,7 +141,7 @@ class Hipay_Gateway_Local_Abstract extends Hipay_Gateway_Abstract
 
             $params = array(
                 "order_id" => $order_id,
-                "paymentProduct" => $this->paymentProduct,
+                "paymentProduct" => Hipay_Helper::getPostData($this->paymentProduct.'-payment_product', $this->paymentProduct),
                 "forceSalesMode" => $this->forceSalesMode(),
                 "deviceFingerprint" => Hipay_Helper::getPostData($this->paymentProduct.'-device_fingerprint'),
                 "phone" => json_decode(Hipay_Helper::getPostData($this->paymentProduct.'-phone'))
@@ -161,8 +164,30 @@ class Hipay_Gateway_Local_Abstract extends Hipay_Gateway_Abstract
         }
     }
 
-    private function forceSalesMode()
+    protected function forceSalesMode()
     {
         return !$this->confHelper->getLocalPayment($this->paymentProduct)["canManualCapture"];
+    }
+
+    /**
+     * Check if it's PayPal v2.
+     *
+     * @return bool
+     * @throws Exception
+     */
+    protected function isPaypalV2()
+    {
+        $paypalOptions = Hipay_Available_Payment::getInstance($this->confHelper)
+            ->getAvailablePaymentProducts('paypal')[0]['options'] ?? [];
+
+        return !empty($paypalOptions['provider_architecture_version'])
+            && $paypalOptions['provider_architecture_version'] === 'v1'
+            && !empty($paypalOptions['payer_id'])
+            && $this->getOperatingMode() == OperatingMode::HOSTED_FIELDS;
+    }
+
+    protected function getOperatingMode()
+    {
+        return $this->confHelper->getPaymentGlobal()['operating_mode'];
     }
 }