+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.5.0
+
+- **Add** : Added Klarna payment method
+- **Add** : Added PayPal v2
+- **Add** : Added new branding colors to HiPay module
+
 ## 2.4.0
 
 - **Add** : Added Mooney reference without redirection